grss: provide a command line tool to search files
grss is a command line tool to search for a pattern in a given file.
This tool is based on the tutorial in the "Command Line Applications in
Rust" guide included in the offical Rust docs.

This package provides two crates: a binary crate for a CLI tool and a
library crate for a library providing the business logic--both called
`grss`.